Part 4 - Exercises 4.15 - 4.23

+++ controllers/blog.js
@@ -1,13 +1,18 @@
+// const jwt = require('jsonwebtoken');
 const blogsRouter = require('express').Router();
 const Blog = require('../models/blog');
+// const User = require('../models/user');
+const middleware = require('../utils/middleware');
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({});
+  const blogs = await Blog.find({}).populate('author', { username: 1, name: 1, id: 1 });
   return response.json(blogs);
 });
 
 blogsRouter.get('/:id', async (request, response) => {
-  const blog = await Blog.findById(request.params.id);
+  console.log('in get :id', request.params);
+  const blog = await Blog.findById(request.params.id).populate('author', { username: 1, name: 1, id: 1 });
+  console.log('blog', blog);
 
   if (!blog) {
     return response.status(404).json({
@@ -18,20 +23,33 @@ blogsRouter.get('/:id', async (request, response) => {
   response.json(blog);
 });
 
-blogsRouter.post('/', async (request, response) => {
+blogsRouter.post('/', middleware.userExtractor, async (request, response) => {
   if (!request.body.title || !request.body.url) {
     return response.status(400).json({
       message: 'Missing data...'
     });
   }
 
-  const blog = new Blog(request.body);
+  const blog = new Blog({
+    title: request.body.title,
+    author: request.author.id,
+    url: request.body.title,
+    likes: request.body.likes,
+  });
+  let savedBlog = await blog.save();
 
-  let result = await blog.save();
-  response.status(201).json(result);
+  request.author.blogs = request.author.blogs.concat(savedBlog._id);
+  await request.author.save();
+
+  response.status(201).json(savedBlog);
 });
 
-blogsRouter.delete('/:id', async (request, response) => {
+blogsRouter.delete('/:id', middleware.userExtractor, async (request, response) => {
+  const blog = await Blog.findById(request.params.id);
+  if (request.author._id.toString() !== blog.author.toString()) {
+    return response.status(401).json({ message: 'action not allowed' });
+  }
+
   await Blog.findByIdAndRemove(request.params.id);
   return response.status(204).end();
 });
